hawkbit-client/rauc-hawkbit-updater: Confirmation implementaion
Initial implementation of Hawkbit's confirmationBase end point. This
version allows to send a request to a custon user backend/software,
which is supposed to implement RaucInstallConfirmation DBus interface
and be capable to make a decision whether an installation of new version
is possible/desired at the moment and accept or decline it.

See
https://eclipse.dev/hawkbit/rest-api/rootcontroller-api-guide.html#_get_tenantcontrollerv1controlleridconfirmationbase
for more into on the API end point.

Signed-off-by: Vyacheslav Yurkov <uvv.mail@gmail.com>